refactor: 기존 UI object의 메서드 OutputView로 이동 및 호출 순서에 맞게 메서드 순서 변경
