+++ src/main/kotlin/view/OutputView.kt
@@ -8,6 +8,11 @@ import domain.TotalPrize
 import domain.WinStatistics
 
 object OutputView {
+
+    fun printRequestMoney() {
+        println("구입금액을 입력해 주세요.")
+    }
+
     fun printPurchasedLottoCount(lottoCount: Int) {
         println("${lottoCount}개를 구매했습니다.")
     }
@@ -17,6 +22,14 @@ object OutputView {
             println(it.lottoNumbers.map { lottoNumber: LottoNumber -> lottoNumber.number })
         }
         println()
+    }
+
+    fun printRequestWinningNumbers() {
+        println("지난 주 당첨 번호를 입력해 주세요.")
+    }
+
+    fun printRequestBonusNumber() {
+        println("보너스 볼을 입력해 주세요.")
     }
 
     fun printWinStatistics(winStatistics: WinStatistics) {
